refactor: reduce redundant code for vendoring
turn impl methods for SrcTar and SrcDir as wrappers for that utility
